Split test 'test_create_user_and_login' into two.
